refactor: Add configurations accessors
