+++ infrastructure/src/main/kotlin/internal/Configurations.kt
@@ -1,0 +1,5 @@
+package com.redmadrobot.build.internal
+
+internal const val api: String = "api"
+internal const val testImplementation: String = "testImplementation"
+internal const val detektPlugins: String = "detektPlugins"
